pyproject.toml
[project]
name = "seed-vault"
version = "1.1.0"
description = "GUI and CLI utility to interface with FDSN seismic data centers to search, download, and archive event and continuous waveform data."
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Physics"
]
readme = "README.md"
authors = [
           {name = "Robert Pickle", email = "robert.pickle@anu.edu.au"},
           {name = "Ben Motevalli", email = "benyamin.motevalli@csiro.au"},
           {name = "Vincent Fazio", email = "vincent.fazio@csiro.au"},
           {name = "Neda Taherifar", email = "neda.taherifar@csiro.au"}
]
keywords = ["seismology", "earthquakes", "seismometers", "FDSN", "ObsPy", "miniseed"] 

[tool.poetry]
package-mode = true
packages = [{ include = "seed_vault" }]

[tool.poetry.scripts]
seed-vault = "seed_vault.cli.cli:cli"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
streamlit = "^1.50"
pandas = "^2.2.2"
folium = "^0.20.0"
requests = "^2.32.3"
streamlit-folium = "0.25.3"
click = "^8.1.7"
tqdm = "^4.67.1"
obspy = "^1.4.1"
tabulate = "^0.9.0"
numpy = "^1.21"
matplotlib = "^3.8.5"
pydantic = "^2.9.1"
streamlit-extras = "^0.7.8"
streamlit-ace = "^0.1.1"
st-pages = "^1.0.1"
python-dotenv = "^1.0.0"
amplitude-analytics = "^1.2.1"
mixpanel = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.10.0"
pytest = "^8.3.4"
sphinx = "^8.1.0"
sphinx-rtd-theme = "^3.0.2"
sphinx-autodoc-typehints = "^2.4.4"
sphinx-material = "^0.0.36"
pytest-cov = "^6.0.0"
coverage-badge = "^1.1.2"

